moved otlp files to separate directory
