+++ pkg/pipeline/pipeline_builder.go
@@ -10,6 +10,7 @@ import (
 	"github.com/netobserv/flowlogs-pipeline/pkg/config"
 	"github.com/netobserv/flowlogs-pipeline/pkg/operational"
 	"github.com/netobserv/flowlogs-pipeline/pkg/pipeline/encode"
+	"github.com/netobserv/flowlogs-pipeline/pkg/pipeline/encode/opentelemetry"
 	"github.com/netobserv/flowlogs-pipeline/pkg/pipeline/extract"
 	"github.com/netobserv/flowlogs-pipeline/pkg/pipeline/extract/conntrack"
 	"github.com/netobserv/flowlogs-pipeline/pkg/pipeline/ingest"
@@ -430,11 +431,11 @@ func getEncoder(opMetrics *operational.Metrics, params config.StageParam) (encod
 	case api.S3Type:
 		encoder, err = encode.NewEncodeS3(opMetrics, params)
 	case api.OtlpLogsType:
-		encoder, err = encode.NewEncodeOtlpLogs(opMetrics, params)
+		encoder, err = opentelemetry.NewEncodeOtlpLogs(opMetrics, params)
 	case api.OtlpMetricsType:
-		encoder, err = encode.NewEncodeOtlpMetrics(opMetrics, params)
+		encoder, err = opentelemetry.NewEncodeOtlpMetrics(opMetrics, params)
 	case api.OtlpTracesType:
-		encoder, err = encode.NewEncodeOtlpTraces(opMetrics, params)
+		encoder, err = opentelemetry.NewEncodeOtlpTraces(opMetrics, params)
 	case api.NoneType:
 		encoder, _ = encode.NewEncodeNone()
 	default:
